requirements.txt
python-telegram-bot==20.7
nest-asyncio==1.6.0